# ggml-cpu: Support Q3_K SIMD on s390x