This pull request aims to include SIMD instruction set for Q3_K quantisation on the s390x platform.

### Before SIMD Benchmark
| model                    | size       | params | backend | threads | test  | t/s           |
|--------------------------|------------|--------|---------|---------|-------|---------------|
| qwen2 1.5B Q3_K - Medium | 780.32 MiB | 1.54 B | BLAS    | 16      | pp512 | 110.51 ± 0.89 |
| qwen2 1.5B Q3_K - Medium | 780.32 MiB | 1.54 B | BLAS    | 16      | tg128 | 8.41 ± 0.22   |
| qwen2 3B Q3_K - Medium   | 1.48 GiB   | 3.09 B | BLAS    | 16      | pp512 | 49.05 ± 0.39  |
| qwen2 3B Q3_K - Medium   | 1.48 GiB   | 3.09 B | BLAS    | 16      | tg128 | 3.37 ± 0.05   |
| qwen2 7B Q3_K - Medium   | 3.54 GiB   | 7.62 B | BLAS    | 16      | pp512 | 24.96 ± 0.12  |
| qwen2 7B Q3_K - Medium   | 3.54 GiB   | 7.62 B | BLAS    | 16      | tg128 | 1.54 ± 0.02   |

### After SIMD Benchmark
| model                    | size       | params | backend | threads | test  | t/s           |
|--------------------------|------------|--------|---------|---------|-------|---------------|
| qwen2 1.5B Q3_K - Medium | 780.32 MiB | 1.54 B | BLAS    | 16      | pp512 | 105.84 ± 2.11 |
| qwen2 1.5B Q3_K - Medium | 780.32 MiB | 1.54 B | BLAS    | 16      | tg128 | 23.93 ± 3.07  |
| qwen2 3B Q3_K - Medium   | 1.48 GiB   | 3.09 B | BLAS    | 16      | pp512 | 55.32 ± 0.20  |
| qwen2 3B Q3_K - Medium   | 1.48 GiB   | 3.09 B | BLAS    | 16      | tg128 | 14.92 ± 0.76  |
| qwen2 7B Q3_K - Medium   | 3.54 GiB   | 7.62 B | BLAS    | 16      | pp512 | 26.98 ± 0.04  |
| qwen2 7B Q3_K - Medium   | 3.54 GiB   | 7.62 B | BLAS    | 16      | tg128 | 6.84 ± 0.09   |

### Verification

To ensure that this implementation did not break anything, the SIMD instruction set has been tested on the following models:
* Tested Qwen2.5 1.5B (Q3_K)
* Tested Qwen2.5 3B (Q3_K)
* Tested Qwen2.5 7B (Q3_K)

> [!NOTE]
> Tests were conducted on an IBM z15 Mainframe with 16 IFLs (cores) and 128 GB Memory on a shared R&D LPAR.

Please review this pull request and consider merging into the main repository. Thank you!